+++ substrate/primitives/runtime/Cargo.toml
@@ -27,6 +27,7 @@ sp-application-crypto = { path = "../application-crypto", default-features = fal
 sp-arithmetic = { path = "../arithmetic", default-features = false}
 sp-core = { path = "../core", default-features = false}
 sp-io = { path = "../io", default-features = false}
+sp-runtime-procedural = { path = "procedural", default-features = false}
 sp-std = { path = "../std", default-features = false}
 sp-weights = { path = "../weights", default-features = false}
 
@@ -56,6 +57,7 @@ std = [
 	"sp-arithmetic/std",
 	"sp-core/std",
 	"sp-io/std",
+	"sp-runtime-procedural/std",
 	"sp-state-machine/std",
 	"sp-std/std",
 	"sp-tracing/std",
