+++ substrate/primitives/runtime/procedural/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "sp-runtime-procedural"
+version = "1.0.0"
+authors.workspace = true
+edition.workspace = true
+license = "Apache-2.0"
+homepage = "https://substrate.io"
+repository.workspace = true
+description = "Runtime Modules shared primitive procedural macros"
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[lib]
+proc-macro = true
+
+[dependencies]
+proc-macro2 = "1.0.56"
+quote = "1.0.28"
+syn = { version = "2.0.38", features = ["full"] }
+
+
+[features]
+default = [ "std" ]
+std = []
